three fma test

+++ test/unit/module/core/three_fma.cpp
@@ -37,12 +37,12 @@ TTS_CASE_WITH("Check behavior of average(wide)",
     using ld_t = long double;
     std::cout << sizeof(ld_t);
     auto [a, e, f] = three_fma(a0, a1, a2);
-    ld_t da = ld_t(a);
-    ld_t de = ld_t(e);
-    ld_t df = ld_t(f);
-    ld_t da0 = ld_t(a0);
-    ld_t da1 = ld_t(a1);
-    ld_t da2 = ld_t(a2);
+    ld_t da = ld_t(double(a));
+    ld_t de = ld_t(double(e));
+    ld_t df = ld_t(double(f));
+    ld_t da0 = ld_t(double(a0));
+    ld_t da1 = ld_t(double(a1));
+    ld_t da2 = ld_t(double(a2));
     TTS_EQUAL(std::fmal(da0, da1, da2), (da+de+df));
   }
 };
